+++ src/test/java/com/example/teamcity/api/BaseApiTest.java
@@ -1,13 +1,17 @@
 package com.example.teamcity.api;
 
 import com.example.teamcity.BaseTest;
+import com.example.teamcity.api.generators.SampleBuildGenerator;
 import com.example.teamcity.api.models.AuthModules;
+import com.example.teamcity.api.models.Build;
 import com.example.teamcity.api.models.ServerAuthSettings;
+import com.example.teamcity.api.requests.CheckedRequests;
 import com.example.teamcity.api.requests.ServerAuthRequest;
 import com.example.teamcity.api.spec.Specifications;
 import org.testng.annotations.AfterSuite;
 import org.testng.annotations.BeforeSuite;
 
+import static com.example.teamcity.api.enums.Endpoint.USERS;
 import static com.example.teamcity.api.generators.TestDataGenerator.generate;
 
 public class BaseApiTest extends BaseTest {
@@ -32,5 +36,20 @@ public class BaseApiTest extends BaseTest {
                 .perProjectPermissions(perProjectPermissions)
                 .modules(authModules)
                 .build());
+    }
+
+    public CheckedRequests setupBuildData() {
+        superUserCheckRequests.getRequest(USERS).create(testData.getUser());
+        return new CheckedRequests(Specifications.authSpec(testData.getUser()));
+    }
+
+    public Build generateBuild(CheckedRequests userCheckRequests) {
+        SampleBuildGenerator sampleBuildGenerator = new SampleBuildGenerator(userCheckRequests);
+        return sampleBuildGenerator.createSampleBuild(testData, true);
+    }
+
+    public void checkBuildResults(Build buildResult) {
+        softy.assertEquals(buildResult.getState(), "finished");
+        softy.assertEquals(buildResult.getStatus(), "SUCCESS");
     }
 }
